Restructure Archive and SerializedFile code into their own libraries
For consistency with the other command implementations - it didn't make much sense that some commands were implemented directly inside UnityDataTool and others as separate libraries.
