+++ Archive/README.md
@@ -1,0 +1,3 @@
+# Archive
+
+See [Documentation/command-archive.md](../Documentation/command-archive.md)
