+++ SerializedFile/README.md
@@ -1,0 +1,3 @@
+# SerializedFile
+
+See [Documentation/command-serialized-file.md](../Documentation/command-serialized-file.md)
